requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
uvicorn==0.27.1